+++ src/core/src/lib/components/formly.field.ts
@@ -1,34 +1,40 @@
 import {
   Component, OnInit, OnChanges, EventEmitter, Input, Output, OnDestroy,
-  ViewContainerRef, ViewChild, ComponentRef, ComponentFactoryResolver, SimpleChanges, DoCheck, AfterContentInit, AfterContentChecked, AfterViewInit, AfterViewChecked,
+  ViewContainerRef, ViewChild, ComponentRef, SimpleChanges, DoCheck, AfterContentInit, AfterContentChecked, AfterViewInit, AfterViewChecked,
 } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { FormlyConfig } from '../services/formly.config';
-import { FieldType } from '../templates/field.type';
 import { FormlyFieldConfig, FormlyFormOptions, FormlyLifeCycleFn, FormlyLifeCycleOptions } from './formly.field.config';
+import { FieldWrapper } from '../templates/field.wrapper';
 
 @Component({
   selector: 'formly-field',
-  template: `
-    <ng-template #fieldComponent></ng-template>
-    <div *ngIf="field.template && !field.fieldGroup" [innerHtml]="field.template"></div>
-  `,
+  template: `<ng-template #container></ng-template>`,
   host: {
     '[style.display]': 'field.hide ? "none":""',
+    '[class]': 'field.className? field.className : className',
   },
 })
 export class FormlyField implements OnInit, OnChanges, DoCheck, AfterContentInit, AfterContentChecked, AfterViewInit, AfterViewChecked, OnDestroy {
+  @Input() field: FormlyFieldConfig;
+  @Input('class') className: string = '';
+
   @Input() set model(m: any) {
     console.warn(`NgxFormly: passing 'model' input to '${this.constructor.name}' component is not required anymore, you may remove it!`);
   }
 
-  @Input() form: FormGroup;
-  @Input() field: FormlyFieldConfig;
-  @Input() options: FormlyFormOptions = {};
-  @Output() modelChange: EventEmitter<any> = new EventEmitter();
-  @ViewChild('fieldComponent', {read: ViewContainerRef}) fieldComponent: ViewContainerRef;
+  @Input() set form(form: FormGroup) {
+    console.warn(`NgxFormly: passing 'form' input to '${this.constructor.name}' component is not required anymore, you may remove it!`);
+  }
 
-  private componentRefs: ComponentRef<FieldType>[] = [];
+  @Input() set options(options: FormlyFormOptions) {
+    console.warn(`NgxFormly: passing 'options' input to '${this.constructor.name}' component is not required anymore, you may remove it!`);
+  }
+
+  @Output() modelChange: EventEmitter<any> = new EventEmitter();
+  @ViewChild('container', {read: ViewContainerRef}) containerRef: ViewContainerRef;
+
+  private componentRefs: ComponentRef<FieldWrapper>[] = [];
 
   constructor(private formlyConfig: FormlyConfig) {}
 
@@ -58,16 +64,12 @@ export class FormlyField implements OnInit, OnChanges, DoCheck, AfterContentInit
 
   ngOnChanges(changes: SimpleChanges) {
     if (changes.field) {
-      this.renderField();
+      this.renderField(this.field, this.containerRef);
     }
 
     this.lifeCycleHooks(this.lifecycle.onChanges);
     this.componentRefs.forEach(ref => {
-      Object.assign(ref.instance, {
-        form: this.form,
-        field: this.field,
-        options: this.options,
-      });
+      Object.assign(ref.instance, { field: this.field });
     });
   }
 
@@ -77,38 +79,18 @@ export class FormlyField implements OnInit, OnChanges, DoCheck, AfterContentInit
     this.componentRefs = [];
   }
 
-  private renderField(): ComponentRef<FieldType> {
+  private renderField(field: FormlyFieldConfig, containerRef: ViewContainerRef) {
     this.componentRefs.forEach(componentRef => componentRef.destroy());
     this.componentRefs = [];
-    if (this.field.template) {
-      return;
-    }
 
-    const type = this.formlyConfig.getType(this.field.type);
+    const wrappers = <any>(field.wrappers || []).map(wrapperName => this.formlyConfig.getWrapper(wrapperName));
+    [...wrappers, { ...this.formlyConfig.getType(field.type), componentFactory: (<any> field)._componentFactory }].forEach(({ componentFactoryResolver, component, componentRef }) => {
+      const ref = componentRef ? componentRef : containerRef.createComponent<FieldWrapper>(componentFactoryResolver.resolveComponentFactory(component));
 
-    let fieldComponent = this.fieldComponent;
-    (this.field.wrappers || []).forEach(wrapperName => {
-      const wrapper = this.formlyConfig.getWrapper(wrapperName);
-      const wrapperRef = this.createComponent(wrapper.componentFactoryResolver, fieldComponent, wrapper.component);
-      fieldComponent = wrapperRef.instance.fieldComponent;
+      Object.assign(ref.instance, { field });
+      this.componentRefs.push(ref);
+      containerRef = ref.instance.fieldComponent;
     });
-
-    return this.createComponent(type.componentFactoryResolver, fieldComponent, type.component);
-  }
-
-  private createComponent(componentFactoryResolver: ComponentFactoryResolver, fieldComponent: ViewContainerRef, component: any): ComponentRef<any> {
-    let componentFactory = componentFactoryResolver.resolveComponentFactory(component);
-    let ref = <ComponentRef<FieldType>>fieldComponent.createComponent(componentFactory);
-
-    Object.assign(ref.instance, {
-        form: this.form,
-        field: this.field,
-        options: this.options,
-    });
-
-    this.componentRefs.push(ref);
-
-    return ref;
   }
 
   private get lifecycle(): FormlyLifeCycleOptions {
@@ -117,7 +99,12 @@ export class FormlyField implements OnInit, OnChanges, DoCheck, AfterContentInit
 
   private lifeCycleHooks(callback: FormlyLifeCycleFn) {
     if (callback) {
-      callback(this.form, this.field, this.field.model, this.options);
+      callback(
+        <FormGroup> (this.field.parent ? this.field.parent.formControl : null),
+        this.field,
+        this.field.model,
+        this.field.options,
+      );
     }
   }
 }