+++ src/core/src/lib/core.ts
@@ -8,3 +8,4 @@ export { FieldType, Field } from './templates/field.type';
 export { FieldArrayType } from './templates/field-array.type';
 export { FieldWrapper } from './templates/field.wrapper';
 export { FormlyModule } from './core.module';
+export { defineHiddenProp as ɵdefineHiddenProp } from './utils';